gzBackend/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.11.2
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.14
django-cors-headers==4.4.0
django-filter==23.5
django-maintenance-mode==0.21.1
django-modelcluster==6.3
django-permissionedforms==0.1
django-redis==5.4.0
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
greenlet==3.0.3
gunicorn==22.0.0
html5lib==1.1
idna==3.7
l18n==2021.3
laces==0.1.1
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pillow_heif==0.17.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fsutil==0.14.1
pytz==2024.1
query==0.1.4
redis==5.0.7
requests==2.32.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wagtail==5.1.1
webencodings==0.5.1
Willow==1.6.3